infra/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pandas-stubs>=2.3.0.250703",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
    "taskipy>=1.14.1",
]


[tool.taskipy.tasks]
dev = "uv run main.py"
lint = "ruff check ."
format = "ruff format ."
test = "pytest"
sort = "isort ."
